+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Implemented enhancements
 
 - GP2-1309 - Remove tooltips
+- GP2-341 - Adaptation target markets documents
 - GP2-1182 - Alignment of elements
 - GP2-1194 - Product finder search page content update and restyle
 - GP2-1130 - Product finder previous result start
@@ -30,6 +31,7 @@
 
 ### Fixed bugs
 
+- GP2-1067 - Fixed LessonCompleted component
 - no-ticket - Prettify, lint and cleanup of product-finder and config boilerplate
 - GP2-1111 - Updates to Stats group.
 - GP2-1188 - Fixed IndexError on Target market research page if country is not selected